Move qfRoundTab from components folder to fix CI/CD
related to  #980
